+++ src/main/java/nextstep/core/theme/ThemeRepository.java
@@ -1,4 +1,4 @@
-package nextstep.core;
+package nextstep.core.theme;
 
 public interface ThemeRepository {
     Theme save(Theme theme);
